+++ packages/kiarina-utils-file/src/kiarina/utils/ext/__init__.py
@@ -32,7 +32,7 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     from ._helpers.detect_extension import detect_extension
     from ._helpers.extract_extension import extract_extension
-    from .settings import settings_manager
+    from ._settings import settings_manager
 
 __version__ = "1.0.0"
 
@@ -40,7 +40,7 @@ __all__ = [
     # .helpers
     "detect_extension",
     "extract_extension",
-    # .settings
+    # ._settings
     "settings_manager",
 ]
 
@@ -55,8 +55,8 @@ def __getattr__(name: str) -> object:
         # .helpers
         "detect_extension": "._helpers.detect_extension",
         "extract_extension": "._helpers.extract_extension",
-        # .settings
-        "settings_manager": ".settings",
+        # ._settings
+        "settings_manager": "._settings",
     }
 
     globals()[name] = getattr(import_module(module_map[name], __name__), name)